app/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
redis==5.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-logstash==0.4.8
google-cloud-aiplatform==1.38.0
aiofiles==23.2.1
requests==2.31.0
beautifulsoup4==4.12.3
